pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "Task Manager – task management system."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dj-database-url>=3.1.2",
    "django>=6.0.3",
    "django-bootstrap5>=26.2",
    "django-filter>=25.2",
    "gunicorn>=25.1.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.2",
]

[dependency-groups]
dev = [
    "ruff>=0.15.5",
]
